+++ objects/obj_leveleditor/Draw_77.gml
@@ -24,6 +24,7 @@ if (quit_level_editor>0)
 			file_delete(working_directory+"/custom_levels/" + string(ds_list_find_value(global.all_loaded_custom_levels,global.select_level_index)) + "/automatic_thumbnail.png")
 		}
 		else
+		if (global.level_name != "")
 		{
 			file_delete(working_directory+"/custom_levels/" + string(global.level_name) + "/automatic_thumbnail.png")
 		}
@@ -40,6 +41,7 @@ if (quit_level_editor>0)
 			sprite_save(thumbnail_sprite,0,working_directory+"/custom_levels/" + string(ds_list_find_value(global.all_loaded_custom_levels,global.select_level_index)) + "/automatic_thumbnail.png");
 		}
 		else
+		if (global.level_name != "")
 		{
 			sprite_save(thumbnail_sprite,0,working_directory+"/custom_levels/" + string(global.level_name) + "/automatic_thumbnail.png");
 		}
@@ -53,11 +55,12 @@ if (quit_level_editor>0)
 		scr_save_custom_level();
 	}
 	#endregion /*Save Level END*/
-
+	
 	if (quit_level_editor>5)
 	{
 		global.create_level_from_template = false; /*Disable global.create_level_from_template after loading everything and quitting the level editor, to let every object have time to load and save*/
 		global.level_name = "";
+		global.select_level_index = 0;
 		room_goto(room_title);
 	}
 	#endregion /*End Game End*/
